Move legacy Setup component to legacy-setup/SetupUsingGCP.
